Fix idempotency issue with python version < 3.5
For earlier versions of Python the dict order was not maintained and
comparing key_usage of a certificate and a new request resulted in a
new certificate request, even if the keys were not changed, only
reordered.

According to section 4.2.1.3 of RFC 5280 [1], the key_usage is a bit
string so, independent of the order the user requested, they will have
the same order in the certificate and certificate sign request.

This patch handle key_usage as a set, fixing the comparision of the
values in the request and the certificate, independent of the order of
the keys, and independent of python's version.

Fixes #45

[1] https://tools.ietf.org/html/rfc5280#page-30

Signed-off-by: Rafael Jeffman <rjeffman@redhat.com>
Signed-off-by: Sergio Oliveira Campos <seocam@redhat.com>