System Roles should consistently use ansible_managed in configuration files it manages
bz#2044640

The certificate role needs to generate the hook scripts with the
correct `ansible_managed` string commented in the same format as
generated by the template module.  Borrowed the method from the
kernel_settings to get the ansible managed comment and updated
the base class in module_utils so that it adds the comment to
the pre and post scripts as follows.

  ==> {pre,post}-scripts/<filename>.sh <==
  #!/bin/bash
  #
  # Ansible managed
  #

  ... script ...

Signed-off-by: Noriko Hosoi <nhosoi@redhat.com>