Let each test use a different certificate file name
This is part of the effort to allow CI tests run in the serialized
manner on one VM for shortening the duration of the CI tests. In
the current tests, the same certificate file path is shared and
there is no cleaning up implemented. When multiple tests run
sequentially, some attribute, e.g., the file ownership, could be
inherited from the previous tests, which could make the test fail
in the attribute verification. To avoid the false negative failure,
this commit chooses different certificate file name per test.