tests: Ensure ipaserver hostname is a FQDN
As certificate system role uses an IPA server to test certmonger
provider, the ipaserver node has to have a FQDN for IPA to be installed.
For the RedHat family of distros, from CentOS/RHEL 7+ and newer,
automatic detection of Ansible's hostname may not select 'systemd'
strategy, failing to set the hostnade FQDN. This failure often happens
when running the tests locally with 'qemu'.

This patch ensure that the 'systemd' strategy is used to set hostname.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>