fix: Workaround getcert issue when cert key-file is missing
When trying to resubmit a certificate request that had the key file
removed, "getcert" hangs indefinitely, causing the role to not respond.

This workaround checks if the private key file has been removed and
fails ressubmiting the certificate request until the situation is fixed,
by the user, using 'getcert stop-tracking --id <cert ID>'.