feat: Support this role in container builds
Feature: Support running the certificate role during container builds.

Reason: This is particularly useful for building bootc derivative OSes.

Result: The role is now officially supported for image mode builds.

In non-booted environments, create a systemd unit that runs the
`getcert` command(s) on first boot, instead of running it directly. We
unfortunately cannot use `ConditionFirstBoot=` for that
(https://github.com/systemd/systemd/issues/8268) nor the
`ConditionFileNotEmpty=!/etc/machine-id` trick, as certmonger cannot run
that early in the boot sequence, and thus even on the first boot
/etc/machine-id is created earlier. So instead conditionalize the unit
on the commands file and remove that after success.

Adjust assert_certificate_parameters.yml to only check for the generated
certificate in non-buildah environments. For buildah, instead check that
the first-boot unit was created and enabled.

Make tests_many_self_signed.yml and tests_fs_attrs full bootc end-to-end
tests. We don't have to do that for all tests, as the mechanics of the
unit generation are fully generic. The most complicated cases are
creating more than one certificate and changing the permissions
afterwards, which these tests capture.

https://issues.redhat.com/browse/RHEL-93207